+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.0 - 2021-09-09
+* Add routes from NyPlugin
+* NyState has new 'get' helpers for TextTheme and MediaQuery
+* NyStorage 'read' method now returns more accurate data types if a type is not specified
+* New template for NyPlugin's
+* Added a MetroService and MetroConsole class for cli operations
+* Pubspec.yaml dependency updates
+
 ## 1.0.0 - 2021-07-07
 
 * Initial release.